refactor(MediaFrame): unify media frame convertors to adaptors
