+++ streamcenter/src/adaptors/mod.rs
@@ -1,1 +1,2 @@
+pub mod flv;
 pub mod rtp;
